# Conflicts:
#	androidApp/build.gradle.kts
#	composeApp/src/desktopMain/kotlin/com/arn/scrobble/main/main.kt
#	extras-common/src/commonMain/kotlin/com/arn/scrobble/billing/BaseBillingRepository.kt
#	extras-nonplay/src/commonMain/kotlin/com/arn/scrobble/billing/BillingRepository.kt
#	extras-play/src/main/java/com/arn/scrobble/billing/BillingRepository.kt
#	fastlane/metadata/android/en-US/full_description.txt
#	gradle/libs.versions.toml

+++ composeApp/src/commonMain/kotlin/com/arn/scrobble/db/ScrobbleSource.kt
@@ -1,8 +1,8 @@
 package com.arn.scrobble.db
 
-import androidx.room.Entity
-import androidx.room.Index
-import androidx.room.PrimaryKey
+import androidx.room3.Entity
+import androidx.room3.Index
+import androidx.room3.PrimaryKey
 import kotlinx.serialization.Serializable
 
 @Entity(
@@ -15,7 +15,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ScrobbleSource(
     @PrimaryKey(autoGenerate = true)
-    val _id: Int = 0,
+    val _id: Long = 0,
     val timeMillis: Long,
     val pkg: String,
 )
